Rename `FilterFactory` back to `Filters`
