+++ client/src/main/java/io/spine/client/TopicBuilder.java
@@ -46,7 +46,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * </pre>
  *
  * @see io.spine.client.TopicFactory#select(Class) to start topic building
- * @see io.spine.client.FilterFactory for filter creation shortcuts
+ * @see Filters for filter creation shortcuts
  * @see AbstractTargetBuilder for more details on this builders API
  */
 public final class TopicBuilder extends AbstractTargetBuilder<Topic, TopicBuilder> {
