+++ client/src/test/java/io/spine/client/TopicBuilderTest.java
@@ -49,12 +49,12 @@ import static io.spine.client.Filter.Operator.EQUAL;
 import static io.spine.client.Filter.Operator.GREATER_OR_EQUAL;
 import static io.spine.client.Filter.Operator.GREATER_THAN;
 import static io.spine.client.Filter.Operator.LESS_OR_EQUAL;
-import static io.spine.client.FilterFactory.all;
-import static io.spine.client.FilterFactory.either;
-import static io.spine.client.FilterFactory.eq;
-import static io.spine.client.FilterFactory.ge;
-import static io.spine.client.FilterFactory.gt;
-import static io.spine.client.FilterFactory.le;
+import static io.spine.client.Filters.all;
+import static io.spine.client.Filters.either;
+import static io.spine.client.Filters.eq;
+import static io.spine.client.Filters.ge;
+import static io.spine.client.Filters.gt;
+import static io.spine.client.Filters.le;
 import static io.spine.client.given.ActorRequestFactoryTestEnv.requestFactory;
 import static io.spine.client.given.TopicBuilderTestEnv.TEST_ENTITY_TYPE;
 import static io.spine.client.given.TopicBuilderTestEnv.TEST_ENTITY_TYPE_URL;
