+++ server/src/test/java/io/spine/server/entity/RecordBasedRepositoryTest.java
@@ -58,8 +58,8 @@ import static com.google.common.collect.Lists.reverse;
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.protobuf.util.FieldMaskUtil.fromFieldNumbers;
 import static io.spine.client.CompositeFilter.CompositeOperator.ALL;
-import static io.spine.client.FilterFactory.all;
-import static io.spine.client.FilterFactory.eq;
+import static io.spine.client.Filters.all;
+import static io.spine.client.Filters.eq;
 import static io.spine.client.OrderBy.Direction.ASCENDING;
 import static io.spine.client.OrderBy.Direction.DESCENDING;
 import static io.spine.protobuf.AnyPacker.pack;
