+++ server/src/test/java/io/spine/server/entity/storage/CompositeQueryParameterTest.java
@@ -35,9 +35,9 @@ import static com.google.common.testing.NullPointerTester.Visibility.PACKAGE;
 import static com.google.common.truth.Truth.assertThat;
 import static io.spine.client.CompositeFilter.CompositeOperator.ALL;
 import static io.spine.client.CompositeFilter.CompositeOperator.CCF_CO_UNDEFINED;
-import static io.spine.client.FilterFactory.eq;
-import static io.spine.client.FilterFactory.ge;
-import static io.spine.client.FilterFactory.lt;
+import static io.spine.client.Filters.eq;
+import static io.spine.client.Filters.ge;
+import static io.spine.client.Filters.lt;
 import static io.spine.server.entity.storage.Columns.findColumn;
 import static io.spine.server.storage.LifecycleFlagField.archived;
 import static io.spine.server.storage.LifecycleFlagField.deleted;
