+++ server/src/test/java/io/spine/server/storage/RecordStorageTest.java
@@ -54,8 +54,8 @@ import static com.google.common.truth.Truth.assertThat;
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.base.Identifier.pack;
 import static io.spine.client.CompositeFilter.CompositeOperator.ALL;
-import static io.spine.client.FilterFactory.all;
-import static io.spine.client.FilterFactory.eq;
+import static io.spine.client.Filters.all;
+import static io.spine.client.Filters.eq;
 import static io.spine.protobuf.AnyPacker.unpack;
 import static io.spine.server.entity.storage.EntityRecordWithColumns.create;
 import static io.spine.server.storage.LifecycleFlagField.archived;
