+++ server/src/test/java/io/spine/system/server/MirrorRepositoryTest.java
@@ -52,7 +52,7 @@ import java.util.Map;
 import static com.google.common.collect.Streams.stream;
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.base.Time.getCurrentTime;
-import static io.spine.client.FilterFactory.eq;
+import static io.spine.client.Filters.eq;
 import static io.spine.protobuf.AnyPacker.pack;
 import static io.spine.protobuf.AnyPacker.unpackFunc;
 import static io.spine.server.storage.LifecycleFlagField.archived;
